Added new logic for Tumbler Puzzle
Turn tumbler and color on that tumbler changed also
    tumbler on different panel with index = index + 1 changed too

Visible panel changed 3 colors
Hidden panel changed 2

I get from master object smartLeds and use setOneLed() function
This function very usefull for small changes in smartLeds array
and it use Rlock on every execution. So need to be carefull with that.
It can inhibit our quest
It may be changed to work with array of smartLeds